Use non-phase compensated version
Also switched the energy measure over to the smaller eigenvalue times
the ratio of eigenvalues.  This seems to give quite good results.  Main
downside is that it still picks up on edges at particular orientations
in quite a nasty fashion.  Also, the precise placement looks a little
haphazard (even if quite close to corners).