Turn ImageBuffer into a class
To make it easier to create, and to somewhat encapsulate the actual
dimensions etc, implement ImageBuffer as a class with a constructor.
This means that you only need to tell it how much padding to use, and
what alignment, and it will sort everything else out.

Alongside this, the buffer filter & image buffer have been split out
into their only library, and are now have their own speed test.  There
is also the start of a verification executable for the filter, which
involved adding a dependency on Eigen.