Updated .cl -> binary mechanism
Previously, use of BUILT_SOURCES provided an effective, but not very
elegant solution.  In particular, it did not maintain the dependency
graph effectively.

The new approach turns the source into a new translation using with xxd,
ap/prepending the namespace as appropriate.  These namespace need to be
individual to the source (in case they get linked together), but this
avoids having to come up with ugly names for the different kernels.