Replaced the old DTCWT implementation
It turned out there was an image rescaling kernel in the old DTCWT's
files as well, so moved that into MiscKernels with its own convenience
library.

Other files needed their header includes updating to use the new
location for the new DTCWT (which previously lived under BufferFilter).

The new DTCWT works, but doesn't produce output in the conventional
location.  This is the next big hurdle.  Consequently, make check fails
interpolation.