Fix display
Difficulties were encountered with the AMD drivers when using events
with OpenGL sync.

It seems these events only complete when you finish() the command queue.
So, we wait for an earlier event to complete, then call finish to wait
for it to definitely conclude.

Also, calling glFinish rather than glFlush is probably wise, in the
absence of a better sync mechanism.

At some stage, could be worth considering splitting out the operation
kernels from the data.