Speed up the triple filter on AMD 7770
The kernel had been observed to be surprisingly slow (~400ms) compared
to the 60ms of a single filter when operating on a small amount of data.
Looks like this was because of the code length.  Reducing slightly, by
sharing the output stage in a loop, gets around this, bringing time down
to closer to 60ms (again).