Simplify DecimateFilterX's kernel
Previously, padding made no difference to the region read in, so that
the accesses to the local memory had to be mangled depending on whether
padding had occured.  This made everything a good deal more complicated
than it should have been, and also take a little longer.  It could have
improved speed a lot on Nvidia cards of even lesser compute capability
(coalesced reads etc.).