Make use of the start() of ImageBuffer
Now most places will respect use of any start location within the image.
There are a few places (e.g. output to two subbands) where these
locations are asserted to be the same.  These will need modifying if the
algorithm is changed so the outputs go into the same buffer.