Fix bug in maximum search
The previous method used a slightly different one to -hessian\grad, and
didn't weight the centre point strongly enough.  The new calculation
fixes both of these.

Also limit the maximum move to at most a single unit along both axis.
Exceeding that suggests something is unstable, so drop the point.  This
seems to get rid of the points appearing outside the image.